This is the configuration used for reproduction results
I basically removed the other two seeds and kept the first one. Also I experimented with the natural dataset as well because it was there, but we did not reported the accuracy in the report. I used the same configuration for the Hierarchical VAE as well with hierarchical attribute = True.